+++ core/common/src/main/kotlin/com/github/jcornaz/miop/FactoryFunctions.kt
@@ -45,9 +45,8 @@ public fun <T> Sequence<T>.openSubscription(context: CoroutineContext = Dispatch
 /**
  * Equivalent of [produce] but starting atomically. (It is guaranteed that [block] is invoked, even if the job is cancelled)
  */
-
 @ExperimentalCoroutinesApi
-internal fun <E> CoroutineScope.produceAtomic(context: CoroutineContext = EmptyCoroutineContext, capacity: Int = 0, block: suspend ProducerScope<E>.() -> Unit): ReceiveChannel<E> {
+public fun <E> CoroutineScope.produceAtomic(context: CoroutineContext = EmptyCoroutineContext, capacity: Int = 0, block: suspend ProducerScope<E>.() -> Unit): ReceiveChannel<E> {
     val result = Channel<E>(capacity)
 
     val job = launch(context, CoroutineStart.ATOMIC) {